move utils.ts->shared/utils.ts, address issues raised in review, new file for command line stuff.
