+++ src/server/main.ts
@@ -38,20 +38,14 @@ import { WindowStateAction } from './reducers/windowStateReducer';
 import * as url from 'url';
 import * as path from 'path';
 import * as log from './log';
-import * as commandLineArgs from 'command-line-args';
 import { Emulator } from './emulator';
 import { WindowManager } from './windowManager';
+import { parseCommandLineArgs } from './commandLineManager'
 
 (process as NodeJS.EventEmitter).on('uncaughtException', (error: Error) => {
     console.error(error);
     log.error('[err-server]', error.message.toString(), JSON.stringify(error.stack));
 });
-
-const optionDefinitions = [
-    { name: 'localstore', type: String },
-];
-
-const options = commandLineArgs(optionDefinitions, { partial: true });
 
 export let mainWindow: Electron.BrowserWindow;
 export let windowManager: WindowManager;
@@ -69,6 +63,8 @@ var onOpenUrl = function (event, url) {
         }
     }
 };
+
+parseCommandLineArgs();
 
 Electron.app.on('will-finish-launching', (event, args) => {
     Electron.ipcMain.on('getUrls', (event, arg) => {
@@ -229,7 +225,6 @@ const createMainWindow = () => {
     mainWindow.loadURL(page);
 }
 
-global["localstore"] = options.localstore || path.join(Electron.app.getPath("userData"), "botframework-emulator");
 Emulator.startup();
 Electron.app.on('ready', createMainWindow);
 Electron.app.on('window-all-closed', function () {
